Using Dar_One in a Grid

Running multiple Dar_One experiments in parallel.

Setting ptracer initial values

PTRACERS_initialFile instead of PTRACERS_ref

initial file is a bin file in the format...

For a 30x30 grid, each tracer will have a 30x30 bin file.

bin files will be placed in the config run folder, under the folder bin_files, and each file should be named something accordingly, i.e. no3_seeds.bin

There will be a handful of default grid builds

30x30

Creating a grid build

Choose the size of your grid, (X, Y).

Update SIZE.h file

set sNx = X and sNy = Y

build

Runtime parameters

data

PARM04

delX=X*1.E0,

delY=Y*1.E0,

Main.DarOneTools.update_delX_delY_for_grid — Function

TODO

Main.DarOneTools.init_tracer_grid — Function

init_tracer_grid(config_obj, tracer_name, init_list, dim, grid_size=(5,3))

Create a bin file of initial values for a tracer. For a grid of size (X, Y), the initial conditions for each grid cell are set in a binary file with X*Y number of floats, in big-endian format, with no headers.

Arguments

config_obj

tracer_name: String

init_list: list of values to initialize the tracer to

dim: "x" or "y" - which dimension to set to the list; values won't change along the other axis

grid_size: tuple ints, default 2x2

Returns

the matrix of initial values

init_tracer_grid(config_obj, tracer_name, init_matrix)

Create a bin file of initial values for a tracer. For a grid of size (X, Y), the initial conditions for each grid cell are set in a binary file with X*Y number of floats, in big-endian format, with no headers.

Arguments

config_obj

tracer_name: String

init_matrix: 2D list of values to write to the bin file

Returns

the matrix of initial values